+++ package.nls.json
@@ -145,8 +145,8 @@
     "AWS.command.cdk.previewStateMachine": "Render state machine graph from CDK application",
     "AWS.command.copyLogStreamName": "Copy Log Stream Name",
     "AWS.command.saveCurrentLogStreamContent": "Save Current Log Content to File",
-    "AWS.command.cwl.changeFilterPattern": "Change filter pattern for search",
-    "AWS.command.cwl.changeTimeFilter": "Change time filter for search",
+    "AWS.command.cwl.changeFilterPattern": "Set/Change filter pattern for search",
+    "AWS.command.cwl.changeTimeFilter": "Set/Change time filter for search",
     "AWS.command.viewLogStream": "View Log Stream...",
     "AWS.command.ssmDocument.createLocalDocument": "Create a new Systems Manager Document locally",
     "AWS.command.ssmDocument.deleteDocument": "Delete Document",
