+++ src/cloudWatchLogs/activation.ts
@@ -17,8 +17,7 @@ import { LogGroupNode } from './explorer/logGroupNode'
 import { LogStreamRegistry } from './registry/logStreamRegistry'
 import { Commands } from '../shared/vscode/commands2'
 import { searchLogGroup } from './commands/searchLogGroup'
-import { changeFilterPattern } from './commands/changeFilterPattern'
-import { changeTimeFilter } from './commands/changeTimeFilter'
+import { changeLogSearchParams } from './changeLogSearch'
 
 export async function activate(context: vscode.ExtensionContext, configuration: Settings): Promise<void> {
     const settings = new CloudWatchLogsSettings(configuration)
@@ -70,8 +69,8 @@ export async function activate(context: vscode.ExtensionContext, configuration: 
 
         Commands.register('aws.cwl.searchLogGroup', async () => await searchLogGroup(registry)),
 
-        Commands.register('aws.cwl.changeFilterPattern', async () => changeFilterPattern(registry)),
+        Commands.register('aws.cwl.changeFilterPattern', async () => changeLogSearchParams(registry, 'filterPattern')),
 
-        Commands.register('aws.cwl.changeTimeFilter', async () => changeTimeFilter(registry))
+        Commands.register('aws.cwl.changeTimeFilter', async () => changeLogSearchParams(registry, 'timeFilter'))
     )
 }
